app/requirements.txt
streamlit>=1.32
databricks-sdk>=0.105,<1
psycopg[binary]>=3.1
